Moved certain processing function from data to process. They were originally written before the process module existed.
